requirements.txt
scipy
numpy
python-dotenv
Jinja2
